+++ ci/script.sh
@@ -13,22 +13,9 @@ main() {
 
     cargo install --path .
 
-    # Test `cross build` / `cross run` for targets that don't have `std` support
+    # `cross build` test for targets where `std` is not available
     case $TARGET in
         thumbv*-none-eabi*)
-            td=$(mktemp -d)
-
-            git clone \
-                --depth 1 \
-                --recursive \
-                https://github.com/rust-lang-nursery/compiler-builtins $td
-
-            pushd $td
-            cross build --features c --target $TARGET
-            popd
-
-            rm -rf $td
-
             td=$(mktemp -d)
 
             git clone \
@@ -46,7 +33,28 @@ main() {
         ;;
     esac
 
-    # Test `cross build` for the other targets
+    # `cross build` test for targets where `std` is not available
+    case $TARGET in
+        thumbv*-none-eabi* | \
+            sparc64-*)
+            td=$(mktemp -d)
+
+            git clone \
+                --depth 1 \
+                --recursive \
+                https://github.com/rust-lang-nursery/compiler-builtins $td
+
+            pushd $td
+            cross build --features c --target $TARGET
+            popd
+
+            rm -rf $td
+
+            return
+            ;;
+    esac
+
+    # `cross build` test for the other targets
     if [ $TARGET = i686-apple-darwin ] || [ $TARGET = i686-unknown-linux-musl ]; then
         td=$(mktemp -d)
 
@@ -69,15 +77,17 @@ main() {
         rm -rf $td
     fi
 
-    # Test `cross test` / `cross run`
+    # `cross test` / `cross run` test for the other targets
     # NOTE(s390x) japaric/cross#3
     # NOTE(*-musl) can't test compiler-builtins because that crate needs
     # cdylibs and musl targets don't support cdylibs
     # NOTE(*-*bsd) no `cross test` support for BSD targets
+    # NOTE(sparc64-*) no `std` available
     case $TARGET in
         i686-unknown-freebsd | \
             i686-unknown-linux-musl | \
             s390x-unknown-linux-gnu | \
+            sparc64-unknown-linux-gnu | \
             x86_64-unknown-freebsd | \
             x86_64-unknown-linux-musl | \
             x86_64-unknown-netbsd)
@@ -114,8 +124,7 @@ main() {
     # Test C++ support
     case $TARGET in
         *-unknown-*bsd | \
-            *-unknown-linux-musl | \
-            thumb*-none-eabi*)
+            *-unknown-linux-musl)
             ;;
         *)
             td=$(mktemp -d)