+++ CHANGELOG.md
@@ -8,14 +8,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - `cross run` support for the thumb targets.
+
 - A `build.xargo` / `target.$TARGET.xargo` option to Cross.toml to use Xargo
   instead of Cargo.
+
+- A `target.$TARGET.image` option to override the Docker image used for
+  `$TARGET`.
 
 ## [v0.1.4] - 2017-01-07
 
 ### Added
 
 - Support for the `arm-unknown-linux-gnueabi` target
+
 - `cross build` support for:
   - `i686-unknown-freebsd`
   - `x86_64-unknown-freebsd`